Cleaned up some odds and ends: - Marked projects as AOT-compatible - Moved some code from `Json` to `VariantJson` for better clarity - Enabled compression tests on Desktop Framework
