+++ src/Apache.Arrow.Compression/Apache.Arrow.Compression.csproj
@@ -1,14 +1,9 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <Description>Provides decompression support for the Arrow IPC format</Description>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(IsWindows)'=='true'">
     <TargetFrameworks>netstandard2.0;net462</TargetFrameworks>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(IsWindows)'!='true'">
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
+    <IsAotCompatible Condition="!$([MSBuild]::IsTargetFrameworkCompatible($(TargetFramework), 'net8.0'))">true</IsAotCompatible>
+    <Description>Provides decompression support for the Arrow IPC format</Description>
   </PropertyGroup>
 
   <ItemGroup>
