+++ src/Apache.Arrow.Flight.Sql/Apache.Arrow.Flight.Sql.csproj
@@ -2,6 +2,7 @@
   <PropertyGroup>
     <TargetFramework>netstandard2.1</TargetFramework>
     <Nullable>enable</Nullable>
+    <IsAotCompatible Condition="!$([MSBuild]::IsTargetFrameworkCompatible($(TargetFramework), 'net8.0'))">true</IsAotCompatible>
   </PropertyGroup>
 
   <ItemGroup>
