+++ src/Apache.Arrow.Flight/Apache.Arrow.Flight.csproj
@@ -1,7 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>netstandard2.0;netstandard2.1;net472</TargetFrameworks>
+    <TargetFrameworks>netstandard2.0;net8.0;net472</TargetFrameworks>
+    <IsAotCompatible Condition="!$([MSBuild]::IsTargetFrameworkCompatible($(TargetFramework), 'net8.0'))">true</IsAotCompatible>
   </PropertyGroup>
 
   <ItemGroup>
