+++ src/Apache.Arrow.Operations/Apache.Arrow.Operations.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <TargetFrameworks>netstandard2.0;net8.0;net462</TargetFrameworks>
-
+    <IsAotCompatible Condition="!$([MSBuild]::IsTargetFrameworkCompatible($(TargetFramework), 'net8.0'))">true</IsAotCompatible>
     <Description>Format-specific codecs for the Apache Arrow Variant binary format, including JSON reader/writer.</Description>
   </PropertyGroup>
