+++ test/Apache.Arrow.Variant.Benchmarks/EncodingBenchmarks.cs
@@ -17,7 +17,7 @@ using System.Buffers;
 using System.Collections.Generic;
 using System.Text;
 using System.Text.Json;
-using Apache.Arrow.Operations.Json;
+using Apache.Arrow.Operations.VariantJson;
 using Apache.Arrow.Scalars.Variant;
 using BenchmarkDotNet.Attributes;
